backend/requirements_free_embeddings.txt
# Free embedding requirements - no OpenAI dependency
sentence-transformers>=2.2.0
torch>=1.9.0
transformers>=4.6.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Existing dependencies
fastapi
uvicorn
supabase
python-dotenv
httpx
pydantic
tqdm